Extracts shared process-orphan tracking into one module
The pidfile helpers and the kill decision — a recorded pid is reaped only when the live process still looks like a JVM, otherwise its stale file is just removed — move from the server manager into a shared module so the instance side reuses the same audited logic instead of duplicating it.
